refactor: move function in chain agnostic permissions
