+++ packages/chain-agnostic-permission/src/index.test.ts
@@ -15,6 +15,9 @@ describe('@metamask/chain-agnostic-permission', () => {
         "getInternalScopesObject",
         "getSessionScopes",
         "getPermittedAccountsForScopes",
+        "getCaipAccountIdsFromScopesObjects",
+        "getAllScopesFromScopesObjects",
+        "isKnownSessionPropertyValue",
         "validateAndNormalizeScopes",
         "bucketScopes",
         "assertIsInternalScopeString",
