+++ packages/chain-agnostic-permission/src/index.ts
@@ -18,6 +18,7 @@ export {
 export {
   getCaipAccountIdsFromScopesObjects,
   getAllScopesFromScopesObjects,
+  isKnownSessionPropertyValue,
 } from './scope/validation';
 export type { Caip25Authorization } from './scope/authorization';
 export {
@@ -31,7 +32,6 @@ export {
   KnownWalletNamespaceRpcMethods,
   KnownNotifications,
   KnownWalletScopeString,
-  isKnownSessionPropertyValue,
 } from './scope/constants';
 export { getSupportedScopeObjects } from './scope/filter';
 export type {
