+++ packages/chain-agnostic-permission/src/scope/constants.ts
@@ -96,17 +96,3 @@ export const KnownNotifications: Record<NonWalletKnownCaipNamespace, string[]> =
 export enum KnownSessionProperties {
   SolanaAccountChangedNotifications = 'solana_accountChanged_notifications',
 }
-
-/**
- * Checks if a given value is a known session property.
- *
- * @param value - The value to check.
- * @returns `true` if the value is a known session property, otherwise `false`.
- */
-export function isKnownSessionPropertyValue(
-  value: string,
-): value is KnownSessionProperties {
-  return Object.values(KnownSessionProperties).includes(
-    value as KnownSessionProperties,
-  );
-}
